+++ pkg/kubectl/BUILD
@@ -14,7 +14,6 @@ go_test(
         "delete_test.go",
         "env_file_test.go",
         "generate_test.go",
-        "kubectl_test.go",
         "namespace_test.go",
         "quota_test.go",
         "resource_filter_test.go",
